Refactored groupBy logic for solve_irls_eqn functions
When applying the solve_irls_eqn equation, the groupBy clause now includes alpha_name in addition to sample_block and label, which simplifies the code logic significantly and should lead to better parallelization.

The cross validation step of both ridge_model and logistic_model has been encapsulated in a helper function to reduce code duplication.

Added alternative logic for input validation for logistic regression, where we don't expect the labels to be centered at 0 with stdev = 1, and we do expect a covariate with stdev = 0 (the intercept).

Signed-off-by: Leland Barnard <leland.barnard@regeneron.com>